Don't use half float for cuda...causing CUDNN_STATUS_BAD_PARAM

+++ config.py
@@ -4,7 +4,7 @@ WINDOW_SIZE: int = 15
 
 if torch.cuda.is_available():
     DEVICE = torch.device('cuda')
-    DTYPE = torch.float16
+    DTYPE = torch.float32
 else:
     DEVICE = torch.device('cpu')
     DTYPE = torch.float32
